+++ source/HAL/interface/HAL/ShaderCompiler.hpp
@@ -17,6 +17,8 @@ namespace HAL {
         Compute     
     };
 
+
+
     struct ShaderCompilerCreateInfo {
         ShaderModel ShaderModelVersion;
         bool        IsDebugMode;
@@ -30,9 +32,9 @@ namespace HAL {
         
         ~ShaderCompiler();
 
-        auto CompileFromString(std::wstring const& data, std::wstring const& entryPoint, ShaderStage target, std::vector<std::wstring> const& defines) const -> std::optional<std::vector<uint32_t>>;
+        auto CompileFromString(std::wstring const& data, std::wstring const& entryPoint, ShaderStage target, std::vector<std::wstring> const& defines) const -> std::optional<std::vector<uint8_t>>;
          
-        auto CompileFromFile(std::wstring const& path, std::wstring const& entryPoint, ShaderStage target, std::vector<std::wstring> const& defines) const -> std::optional<std::vector<uint32_t>>;
+        auto CompileFromFile(std::wstring const& path, std::wstring const& entryPoint, ShaderStage target, std::vector<std::wstring> const& defines) const -> std::optional<std::vector<uint8_t>>;
              
     private:    
         InternalPtr<Internal, InternalSize_ShaderCompiler> m_pInternal;
